Add explicit type casts for RECT initialization and include windef.h header

+++ Source Main 5.2/source/MUMain/Winmain.cpp
@@ -1137,7 +1137,7 @@ int APIENTRY WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, PSTR szCmdLin
 
     if (g_bUseWindowMode == TRUE)
     {
-        RECT rc = { 0, 0, WindowWidth, WindowHeight };
+        RECT rc = { 0, 0, static_cast<LONG>(WindowWidth), static_cast<LONG>(WindowHeight) };
         AdjustWindowRect(&rc, WS_OVERLAPPED | WS_CAPTION | WS_SYSMENU | WS_MINIMIZEBOX | WS_BORDER | WS_CLIPCHILDREN, NULL);
         g_hWnd = CreateWindow(
             windowName, windowName,
